Made changes to database helper functions

+++ ExpenseTracker/backend/db_interaction.py
@@ -5,7 +5,8 @@ import pymysql
 from dotenv import load_dotenv
 import os
 from contextlib import contextmanager
-from logging_setup import setup_logger
+from ExpenseTracker.backend.logging_setup import setup_logger
+from datetime import date
 
 # loads .env file
 load_dotenv()
@@ -14,7 +15,7 @@ load_dotenv()
 logger = setup_logger("db_interaction.log", "db_interaction.log", "INFO")
 
 @contextmanager
-def get_db_cursor():
+def get_db_cursor(to_be_commited=False):
     '''
     Function to obtain database cursor
     :return:
@@ -36,15 +37,18 @@ def get_db_cursor():
         cursor = mysql_connection.cursor()
         print("Connected to MySQL database")
         yield cursor
-
-        # close cursor
-        cursor.close()
-        # close connection
-        mysql_connection.close()
+        # In case of any change made to the database the change is temporarily stored unless commited.
+        if to_be_commited:
+            mysql_connection.commit()
 
     except Exception as e:
         mysql_connection.rollback()
         raise e
+    finally:
+        # close cursor
+        cursor.close()
+        # close connection
+        mysql_connection.close()
 
 def fetch_all_records():
     '''
@@ -60,9 +64,11 @@ def fetch_all_records():
         expenses = cursor.fetchall()
         return expenses
 
-def fetch_expenses_for_date(expense_date):
+def fetch_expenses_for_date(expense_date: date, userid : int):
     '''
-    Function to fetch all expenses for date
+    Function to fetch all expenses for a specific user and date
+    :param expense_date: Expense date
+    :param userid: User ID
     :return:
     '''
     logger.info('Fetching all expenses for date')
@@ -70,48 +76,50 @@ def fetch_expenses_for_date(expense_date):
     # fetch db cursor
     with get_db_cursor() as cursor:
         # fetch all expenses
-        cursor.execute("select * from expenses where expense_date = %s;", (expense_date,))
+        cursor.execute("select * from expenses where expense_date = %s and id = %s;", (expense_date,userid))
         expenses = cursor.fetchall()
         return expenses
 
-def insert_into_database(expense_date,amt,cat,notes):
+def insert_into_database(expense_date,userid,amt,cat,notes):
     '''
     Function to insert entry into database
     :param expense_date: Expense date
+    :param userid: User ID
     :param amt: Amount spent
     :param cat: Category the amount was spent in viz. food, clothing etc.
     :param notes: Description of the expense
     :return:
     '''
-    logger.info('Inserting data into database')
+    logger.info(f'Inserting data into database corresponding to expense date {expense_date} and user {userid}')
 
-    with get_db_cursor() as cursor:
-        cursor.execute("insert into expenses (expense_date, amount, category, notes) values (%s, %s, %s, %s);", (expense_date,amt,cat,notes,))
+    with get_db_cursor(to_be_commited=True) as cursor:
+        cursor.execute("insert into expenses (id ,expense_date, amount, category, notes) values (%s, %s, %s, %s, %s);", (userid,expense_date,amt,cat,notes))
         # In case of any change made to the database the change is temporarily stored unless commited.
-        cursor._connection.commit()
+        # cursor._connection.commit()
 
-def delete_records_from_database_for_a_date(expense_date):
+def delete_records_from_database_for_a_date(expense_date,userid):
     '''
     Function to delete from database entries related to expense date.
     :param expense_date: Remove a record from database with expense date.
     :return:
     '''
-    logger.info('Deleting data from database')
+    logger.info(f'Deleting data from database corresponding to expense date {expense_date} and user {userid}')
 
-    with get_db_cursor() as cursor:
-        cursor.execute("delete from expenses where expense_date = %s;", (expense_date,))
-        cursor._connection.commit()
+    with get_db_cursor(to_be_commited=True) as cursor:
+        cursor.execute("delete from expenses where expense_date = %s and id = %s;", (expense_date,userid))
+        # cursor._connection.commit()
 
-def reset_database():
+def reset_database(user_id: int):
     '''
-    Function to reset database with resetting the auto increment which is the ID i.e. remove all entries from database such that new entry added starts with id = 1.
+    Function to remove all expense entries from the database corresponding to a specific user.
+    :param user_id: User ID
     :return:
     '''
     logger.info('Resetting database')
 
-    with get_db_cursor() as cursor:
-        cursor.execute("truncate table expenses;")
-        cursor._connection.commit()
+    with get_db_cursor(to_be_commited=True) as cursor:
+        cursor.execute("delete from expenses where id = %s;", (user_id,))
+        # cursor._connection.commit()
 
 
 def fetch_expenses_summary(expense_date1,expense_date2,userid : int):
@@ -140,9 +148,9 @@ def register_user(username,password):
 
     logger.info('Inserting new user info into database')
 
-    with get_db_cursor() as cursor:
+    with get_db_cursor(to_be_commited=True) as cursor:
         cursor.execute('insert into LOGGED_USERS (USERNAME, PASSWORD) values (%s, %s);', (username, password,))
-        cursor._connection.commit()
+        # cursor._connection.commit()
 
 def check_for_logged_user(username,password):
     '''
@@ -161,7 +169,7 @@ def check_for_logged_user(username,password):
 if __name__ == '__main__':
     print(fetch_all_records())
     # fetch_expenses_for_date('2024-08-02')
-    # insert_into_database('2025-01-01',5000.0,'Shopping','Purchased apparels')
+    insert_into_database('2025-01-01',5000.0,'Shopping','Purchased apparels')
     # delete_from_database("2025-01-01")
     # fetch_expenses_for_date('2025-01-01')
     # fetch_expenses_categorywise_between_dates("2024-08-02","2024-12-31")
